requirements.txt
altgraph==0.17.4
defusedxml==0.7.1
fonttools==4.53.1
ImageHash==4.3.1
imageio==2.35.1
imutils==0.5.4
lazy_loader==0.4
networkx==3.3
numpy==1.26.4
odfpy==1.4.1
opencv-python==4.10.0.84
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
potracer==0.0.4
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pytesseract==0.3.10
PyWavelets==1.6.0
pywin32-ctypes==0.2.2
scikit-image==0.24.0
scipy==1.14.0
svg.path==6.3
tesseract==0.1.3
tifffile==2024.9.20